multi-asset-portfolio/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multi-asset-portfolio"
version = "0.1.0"
description = "Automated verification and dynamic allocation system for multi-asset portfolios"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Multi-Agent Shogun Team" }
]
keywords = [
    "portfolio",
    "allocation",
    "trading",
    "finance",
    "risk-parity",
    "hrp",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Typing :: Typed",
]

dependencies = [
    # Core
    "numpy>=1.24.0,<2.4",  # Numba requires NumPy < 2.4
    "pandas>=2.0.0",
    "polars>=0.19.0",

    # Numba JIT acceleration
    "numba>=0.60.0",

    # Data Sources
    "yfinance>=0.2.28",
    "ccxt>=4.0.0",

    # Scientific Computing
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "scikit-optimize>=0.9.0",  # Bayesian optimization

    # Configuration & Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",

    # Logging
    "structlog>=23.1.0",

    # Storage
    "duckdb>=0.9.0",
    "pyarrow>=14.0.0",

    # Memory Profiling
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-xdist>=3.3.0",
    "hypothesis>=6.82.0",

    # Type Checking
    "mypy>=1.5.0",
    "types-PyYAML>=6.0.0",

    # Linting & Formatting
    "ruff>=0.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "vulture>=2.10",

    # Pre-commit
    "pre-commit>=3.3.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.2.0",
    "mkdocstrings[python]>=0.22.0",
]

all = [
    "multi-asset-portfolio[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/example/multi-asset-portfolio"
Documentation = "https://github.com/example/multi-asset-portfolio#readme"
Repository = "https://github.com/example/multi-asset-portfolio"
Issues = "https://github.com/example/multi-asset-portfolio/issues"

[project.scripts]
portfolio = "src.main:main"

# =============================================================================
# Tool Configurations
# =============================================================================

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "-v",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "architecture: marks tests as architecture/design rule tests (QA-006)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true
show_error_codes = true
show_column_numbers = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "yfinance.*",
    "ccxt.*",
    "duckdb.*",
    "scipy.*",
    "sklearn.*",
    "skopt.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument
    "ARG002", # unused method argument
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG", "S101", "PLR2004"]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["src"]
force-single-line = false
lines-after-imports = 2

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]
skip_gitignore = true

[tool.vulture]
min_confidence = 80
paths = ["src"]
exclude = [
    "*/__pycache__",
    "*.pyc",
    ".git",
    ".venv",
    "venv",
]
ignore_names = [
    "_*",
    "test_*",
    "Test*",
    "setup",
    "teardown",
]

# =============================================================================
# Hatch Configuration
# =============================================================================

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.envs.default]
dependencies = [
    "multi-asset-portfolio[dev]",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "pytest --cov=src --cov-report=term-missing --cov-report=html {args:tests}"
test-unit = "pytest tests/unit {args}"
test-integration = "pytest tests/integration {args}"
test-architecture = "pytest tests/architecture -m architecture {args}"
type-check = "mypy src tests"
lint = "ruff check src tests"
format = "black src tests && isort src tests"
all = ["format", "lint", "type-check", "test"]

[tool.hatch.envs.docs]
dependencies = [
    "multi-asset-portfolio[docs]",
]

[tool.hatch.envs.docs.scripts]
build = "mkdocs build"
serve = "mkdocs serve"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
